Converted more vars to Value class, fixed odometer

+++ CANserver-master/CANserver-master/CANserver/carDataVariables.h
@@ -5,7 +5,7 @@
 //
 
 static double nominalEnergyRemaining = 0.0;
-static int expectedEnergyRemaining = 0;
+static int expectedEnergyRemaining = 0.0;
 static double nominalFullPackEnergy = 0.0;
 
 static double frontInverterTemp = 0.0;
@@ -15,14 +15,14 @@ static double minBattTemp = 0.0;
 static double maxBattTemp = 0.0;
 static double avgBattTemp = 0.0;
 
-static int cabin_humidity = 0;
-static int cabin_temp = 0;
+static int cabin_humidity = 0.0;
+static int cabin_temp = 0.0;
 
-static double maxRegen = 0;
-static double maxDischarge = 0;
+static double maxRegen = 0.0;
+static double maxDischarge = 0.0;
 
-static double frontPower = 0;
-static double rearPower = 0;
+static double frontPower = 0.0;
+static double rearPower = 0.0;
 static double frontPowerLimit = 0;
 static double rearPowerLimit = 0;
 
@@ -33,18 +33,18 @@ static double battPower = 0.0;
 static int frontTorque = 0;
 static int rearTorque = 0;
 
-static double socAVE = 0;
-static double battTempPct = 0;
+static double socAVE = 0.0;
+static double battTempPct = 0.0;
 
 static double coolantFlowBatActual = 0.0;
 static double coolantFlowPTActual = 0.0;
 
-static int tempCoolandBatInlet = 0;
-static int tempCoolandBatPTlet = 0;
+static double tempCoolandBatInlet = 0;
+static double tempCoolandBatPTlet = 0;
 
-static int chargeLineVoltage = 0;
-static int chargeLineCurrent = 0;
-static int chargeLinePower = 0;
+static double chargeLineVoltage = 0;
+static double chargeLineCurrent = 0;
+static double chargeLinePower = 0;
 
 static int displayOn = 0;
